requirements.txt
GitPython==0.3.2.RC1
pbr
PyYAML
giturlparse.py>=0.0.5
